Student-Projects/group10/requirements.txt
modelzoo-client[transformers]
googleapiclient.discovery
pandas
transformers
emoji
re